Move integration creation helpers to own file
